Refactor some utility methods
Allows them to be more universally used
